refactor(code-interpreter): one export per file for session-sandbox utils
Split session-sandbox-gc.util.ts into release-session-sandboxes.util.ts and sweep-expired-session-sandboxes.util.ts, and extract the shared SESSION_SANDBOX_METADATA_KEY into its own constant file (so get-or-create is single-export too). Specs split to match.
