+++ packages/twenty-server/src/engine/core-modules/code-interpreter/constants/session-sandbox-metadata-key.constant.ts
@@ -1,0 +1,1 @@
+export const SESSION_SANDBOX_METADATA_KEY = 'twentySessionId';
